% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_geneSim.R
\name{cal_geneSim}
\alias{cal_geneSim}
\title{Calculate the functional similarities between two genes based on Gene Ontology.}
\usage{
cal_geneSim(
  gene1,
  gene2,
  geneID2GO = KYgeneID2GO,
  simFun = c("Resnik", "Lin", "JiangConrath", "relevance"),
  geneFun = c("mean", "max", "simAvg", "simMax", "simMin"),
  category = "BP",
  return = F
)
}
\arguments{
\item{gene1}{The first gene, such as "KY.Chr1.1".}

\item{gene2}{The second gene, such as "KY.Chr1.10".}

\item{geneID2GO}{A list contains gene IDs and GO terms.}

\item{simFun}{The calculation method between pairwise GO terms.}

\item{geneFun}{The calculation method between pairwise genes.}

\item{category}{Gene Ontology categories. Cellular component as CC; biological process as BP; molecular function as MF.}

\item{return}{This function will stop by default when error happens. If return is TRUE, then it will return NA.}
}
\value{
A value of functional similarity.
}
\description{
Returns the pairwise similarities between two genes. Different calculation geneFun are provided as implemented in GOSim R package.
}
\examples{
cal_geneSim("KY.Chr1.1","KY.Chr1.1000")
}
\author{
Chen Zaohuang
}
