% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneOrientation.R
\name{geneOrientation}
\alias{geneOrientation}
\title{Identify gene organization}
\usage{
geneOrientation(
  data,
  Orientation = c("HH", "HT", "TT"),
  maxDistance = 1000,
  minDistance = 1,
  size = 2
)
}
\arguments{
\item{data}{A data.frame contains colnumn names including "Geneid","Chr","Start", "End", "Strand" in order, others columns can exist but are not necessary.}

\item{Orientation}{The orientation of neighboring genes. It could be HH, HT or TT.}
}
\value{
A list that contains gene orientation and genes
}
\description{
Identify gene organization
}
\examples{
ci.H2H<-geneOrientation(HT.readcount,Orientation ="HH" ,maxDistance = 1000,minDistance=1)
}
\author{
Chen Zaohuang
}
