% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_geneSim.R
\name{get_GO}
\alias{get_GO}
\title{Get GO terms}
\usage{
get_GO(geneID, geneID2GO = KYgeneID2GO, remove = T)
}
\arguments{
\item{geneID}{Gene ID such as "KY.Chr1.1".}

\item{geneID2GO}{A list contains gene IDs and GO terms.}

\item{remove}{Keep gene ID or not. If True, then remove the geneID.}
}
\value{
A vector of GO terms.
}
\description{
Returns GO terms based the gene ID provided.
}
\examples{
test<-get_GO("KY.Chr1.1",geneID2GO=KYgeneID2GO)
}
\author{
Chen Zaohuang
}
